backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Caching
redis==5.0.1
hiredis==2.2.3

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1

# Selenium (for JavaScript-heavy sites)
selenium==4.15.2
webdriver-manager==4.0.1

# AI/ML
google-generativeai==0.3.1
groq>=0.11.1

# Text Analysis
nltk==3.8.1
textstat==0.7.3

# Authentication
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
validators==0.22.0
loguru==0.7.2
pydantic==2.5.0
